Revert "[Overview][Serve] Add Recent Serve Applications Card (#34642)"
This reverts commit 12c97241f2922dd0f487e456da43e185b8303d8f.
