+++ dashboard/client/src/pages/overview/OverviewPage.tsx
@@ -1,19 +1,11 @@
-import { createStyles, Grid, makeStyles } from "@material-ui/core";
-import classNames from "classnames";
+import { createStyles, makeStyles } from "@material-ui/core";
 import React from "react";
 import { CollapsibleSection } from "../../common/CollapsibleSection";
-import {
-  NodeStatusCard,
-  ResourceStatusCard,
-} from "../../components/AutoscalerStatusCards";
 import EventTable from "../../components/EventTable";
-import { useRayStatus } from "../job/hook/useClusterStatus";
 import { MainNavPageInfo } from "../layout/mainNavContext";
 import { ClusterUtilizationCard } from "./cards/ClusterUtilizationCard";
 import { NodeCountCard } from "./cards/NodeCountCard";
-import { OverviewCard } from "./cards/OverviewCard";
 import { RecentJobsCard } from "./cards/RecentJobsCard";
-import { RecentServeCard } from "./cards/RecentServeCard";
 
 const useStyles = makeStyles((theme) =>
   createStyles({
@@ -48,8 +40,6 @@ const useStyles = makeStyles((theme) =>
 export const OverviewPage = () => {
   const classes = useStyles();
 
-  const { cluster_status } = useRayStatus();
-
   return (
     <div className={classes.root}>
       <MainNavPageInfo
@@ -57,31 +47,9 @@ export const OverviewPage = () => {
       />
       <div className={classes.overviewCardsContainer}>
         <ClusterUtilizationCard className={classes.overviewCard} />
+        <NodeCountCard className={classes.overviewCard} />
         <RecentJobsCard className={classes.overviewCard} />
-        <RecentServeCard className={classes.overviewCard} />
       </div>
-
-      <CollapsibleSection
-        className={classes.section}
-        title="Cluster status and autoscaler"
-        startExpanded
-      >
-        {
-          <div className={classes.overviewCardsContainer}>
-            <NodeCountCard className={classes.overviewCard} />
-            <OverviewCard
-              className={classNames(classes.root, classes.overviewCard)}
-            >
-              <NodeStatusCard cluster_status={cluster_status} />
-            </OverviewCard>
-            <OverviewCard
-              className={classNames(classes.root, classes.overviewCard)}
-            >
-              <ResourceStatusCard cluster_status={cluster_status} />
-            </OverviewCard>
-          </div>
-        }
-      </CollapsibleSection>
 
       <CollapsibleSection
         className={classes.section}
