+++ dashboard/client/src/pages/serve/ServeApplicationDetailPage.tsx
@@ -111,7 +111,7 @@ export const ServeApplicationDetailPage = () => {
           },
           {
             label: "Application config",
-            content: application.deployed_app_config ? (
+            content: (
               <CodeDialogButton
                 title={
                   application.name
@@ -120,8 +120,6 @@ export const ServeApplicationDetailPage = () => {
                 }
                 code={application.deployed_app_config}
               />
-            ) : (
-              <Typography>-</Typography>
             ),
           },
           {
@@ -139,12 +137,6 @@ export const ServeApplicationDetailPage = () => {
                 startTime={application.last_deployed_time_s * 1000}
               />
             ),
-          },
-          {
-            label: "Import path",
-            content: {
-              value: application?.deployed_app_config?.import_path || "-",
-            },
           },
         ]}
       />
